+++ kt/api-generator/src/main/kotlin/godot/codegen/services/impl/GenerationService.kt
@@ -846,7 +846,7 @@ class GenerationService(
                 val appliedDefault = if ((argument.isEnum() || argument.isBitField()) && defaultValueKotlinCode != null) {
                     enumService.findEnumValue(
                         argumentTypeClassName,
-                        defaultValueKotlinCode.toInt()
+                        defaultValueKotlinCode.toLong()
                     ).name
                 } else {
                     defaultValueKotlinCode
